Add edit user functionality
Many changes had to be made to enable this functionality. The edit user view and save user controller method now work with either a new user or an existing user. Some changes in the model were necessary, like removing the password validation via schema, so that it is possible now that the password does not need to be changed when editing a user.